frontend: Use router-view for App.vue, move contents to HomeView
