+++ frontend/src/router/index.ts
@@ -1,12 +1,12 @@
 // router/index.ts
 import { createRouter, createWebHistory } from 'vue-router'
-import App from '../App.vue'
+import HomeView from '../views/HomeView.vue'
 
 const routes = [
     {
         path: '/',
-        name: 'App',
-        component: App,
+        name: 'home',
+        component: HomeView
     },
 ]
